requirements.txt
Django>=1.10.5
pytz>=2016.10
django-mptt>=0.8.7
django-taggit>=0.22.0
django-meta>=1.3.2
django-ckeditor>=5.2.1
pillow>=4.0.0